% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytotype_repro_mate.R
\name{cytotype_repro_mate}
\alias{cytotype_repro_mate}
\title{Reproduction function.}
\usage{
cytotype_repro_mate(cgen, b, cc, gnum.vec, s, mc)
}
\arguments{
\item{cgen}{Number of mature individuals of each cytotype in the current generation.
Must be a list of three numeric values. For example, \code{cgen = c(100, 100, 100)}.}

\item{b}{Proportion of unreduced gamete formed by each diploid and tetraploid individual.
Must be a single integer between 0 and 1.}

\item{cc}{Proportion of 3n gamete formation from each triploid individual.
Must be a single integer between 0 and 1.}

\item{gnum.vec}{The mean number of gametes produced by each individual for each cytotype.
Must be a list of three numeric values. For example, \code{gnum.vec = c(100, 100, 100)}.}

\item{s}{Selfing rate. Must be a single integer between 0 and 1.}

\item{mc}{Strength of mating choice. Must be a single integer between 0 and 1.}
}
\value{
Output is a generation log including number of 2x offspring, 3x offspring, 4x offspring, gametes sampled per 2x individual, gametes sampled per 3x individual,
and gametes sampled per 4x individual.
}
\description{
This is a function to generate the number of offspring produced per generation.
Assumes a mixed mating system where both selfing and outcrossing occur.
Also allows group-based assortative mating (see Otto et al. 2008).
This function prevents selfing from occurring during outcrossing.
}
\references{
Otto, S. P., Servedio, M. R., Nuismer, S. L. (2008). Frequency-dependent selection and the evolution of assortative mating. \emph{Genetics}, 179(4) 2091 - 2112. doi: 10.1534/genetics.107.084418
}
