% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.copula.R
\name{env.copula}
\alias{env.copula}
\title{Environmental Copula -  Calculates correlation probability.}
\usage{
env.copula(rho, generations)
}
\arguments{
\item{rho}{Correlation coefficient. Must be a single integer between 0 and 1.}

\item{generations}{Number of generations to simulate. Must be a numeric value.}
}
\value{
A list of two matrix: X and U. Here, X is the matrix which was simulated from
a multivariate normal distribution. The U matrix is simply the X matrix transformed into a uniform distribution.
Each matrix has 3 columns, one for each cytotype. The length of the matrix matches the number of generations provided.
}
\description{
Calculates the probability coefficient which allow cytotype and stage correlations
}
