% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.iter.R
\name{format.iter}
\alias{format.iter}
\title{Format iteration output}
\usage{
\method{format}{iter}(popvect, generations)
}
\arguments{
\item{popvect}{Output from gen-iter-f-choosy function.}

\item{generations}{Number of generations included.}
}
\value{
Output file is a data frame where each row is a generation. The columns are as follows,
\itemize{
\item V1: number of immature diploids.
\item V2: number of immature triploids.
\item V3: number of immature tetraploids.
\item V4: number of mature diploids.
\item V5: number of mature triploids.
\item V6: number of mature tetraploids.
\item V7: number of diploid offspring produced during t - 1.
\item V8: number of triploid offspring produced during t - 1.
\item V9: number of tetraploid offspring produced during t - 1.
\item V10: number of gametes per diploid individual at t - 1.
\item V11: number of gametes per triploid individual at t - 1.
\item V12: number of gametes per tetraploid individual at t - 1.
\item gen: generation.
\item sum: total number of individuals.
\item sum2x-sum4x: total number of each cytotype at that generation.
\item V1a:V6a: relative abundance of V1:V6 in that generation.
\item C2 is the relative abundance of all diploids (ie. sum2x/sum).
\item C3 is the relative abundance of all triploids (ie. sum3x/sum).
\item C4 is the relative abundance of all tetraploids (ie. sum4x/sum)
}
}
\description{
This function formats the output of gen-iter-f-choosy.
}
