% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.iter.f.choosy.R
\name{gen.iter.f.choosy}
\alias{gen.iter.f.choosy}
\title{Simulate multiple generations}
\usage{
gen.iter.f.choosy(
  generations,
  init.pop,
  env.ci,
  aii.vec,
  as.matur,
  as.msurv,
  d,
  gnum.base,
  b,
  cc,
  s,
  mc,
  density.type = "all",
  mate.lazy = FALSE
)
}
\arguments{
\item{generations}{Number of generations to iterate.}

\item{init.pop}{Initial population size, ie. number of mature diploids in founding population.}

\item{env.ci}{Proportion of environmental variance used to define mature survival rate per generation
with a normal distribution where the mean is the previous generations maturation rate.}

\item{aii.vec}{The probability of an immature individual surviving (cytotype-specific).}

\item{as.matur}{Probability of maturation from immature to mature stage (list).}

\item{as.msurv}{probability of mature survival (list).}

\item{d}{Strength of density dependency on gamete production.}

\item{gnum.base}{List of three numeric values, represent the base level mean
number of gametes per individual per cytotype. We defined gametes per
generation by a poisson distribution where lambda is equal to the this value
times the exponential of the product of strength of density dependency (ie. d)
and the sum of all individuals at time t or all individuals of a certain cytotype
at time t. Sum type is set by density type.}

\item{b}{Proportion of unreduced gamete formed by each diploid and tetraploid individual.}

\item{cc}{Proportion of 3n gamete formation from each triploid individual.}

\item{s}{Selfing rate.}

\item{mc}{Strength of mating choice.}

\item{density.type}{"all" sets gamete per generation sum as the sum of all
individuals at time t. "like-cytotype" sets gamete per generation sum as
the sum of both immature and mature individuals of the cytotype at time t.
Defaults to "all".}

\item{mate.lazy}{Set to be FALSE, unless set to TRUE. FALSE prevents selfing from occurring during outcrossing. However, this increases the computational time by 31x!}
}
\value{
A single data frame as defined by \code{format.iter()}.
}
\description{
This is a function that loops through a set number of generations
to look at population dynamics overtime.
}
