% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.cls.R
\name{mars.cls}
\alias{mars.cls}
\title{MAR(1) Conditional least squares parameter estimates}
\usage{
mars.cls(comm.mat, covariate = NULL)
}
\arguments{
\item{comm.mat}{Data frame with log-scale abundance for each time step following burn-in.
We expect one column per species and one row per time-step.}

\item{covariate}{Matrix of covariates (ex. precipitation per time step) with each column representing a variable and
each row representing a time step. Defaults to "NULL" when no matrix is supplied.}
}
\value{
Resulting data frame includes:
\itemize{
\item \strong{A}, vector of length p (p = number of species) containing the estimates of the intrinsic rate of natural increase of each species.
\item \strong{B},  a matrix of p x p estimates where the diagonal elements represent the intra-specific, density-dependent effects.
The elements \eqn{b_{ij}} gives the effect of the abundance of species j on per capita growth rate of species i.
\item \strong{sigma}, a matrix of p x p, represents environmental noise variance-covariance matrix.
\item \strong{C}, a vector of estimated coefficients for every covariate representing the effects of every covariate on every species.
\item \strong{E}, vector representing stochastic environmental variability.
\item \strong{Yhat}, \eqn{\hat{Y}} estimator of predicted abundance.
\item \strong{R2}, \eqn{R^2}, proportion of explained variation in the log scale population abundance for each species.
\item \strong{R2_D}, conditional \eqn{R^2}, proportion of variation in the change in one unit of time of the log scale population
abundance explained by the model for each species.
\item \strong{AIC}, Akaike information criterion.
\item \strong{BIC}, Bayesian information criterion.
\item \strong{lnlike}, maximized log likelihood of the MAR(1) model.
}
}
\description{
As described in  Ives et al. 2003, the Multivariate Autoregressive Model, also known as the MAR(1) model,
is a discrete-time model for a multispecies stochastic community subject to environmental noise.  It is a Markov process.
Given multispecies time series (log) abundance data without observation error, parameter estimation of this model parameters
can be quickly done via Conditional Least Squares.
}
\references{
Ives, A. R., Dennis, B., Cottingham, K. L., Carpenter, S. R. (2003). Estimating community stability and ecological interactions from time-series data.
\emph{Ecological Monographs}, 72(2). doi:10.1890/0012-9615(2003)073\link{0301:ECSAEI}2.0.CO;2
}
