% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mature.surv.calc.R
\name{mature.surv.calc}
\alias{mature.surv.calc}
\title{Mature survival calculation.}
\usage{
mature.surv.calc(env.ci, as.msurv)
}
\arguments{
\item{env.ci}{Proportion of environmental variance used to define mature survival rate per generation
with a beta distribution. This number must be in between 0 and 1, but cannot be equal to 0 or 1.}

\item{as.msurv}{Mean probability of mature survival (list).}
}
\value{
Probability of mature survival.
}
\description{
Calculates the probability of mature survival based on a beta binomial distribution.
}
