% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu.options.R
\name{mu.options}
\alias{mu.options}
\title{Max mean allowed}
\usage{
mu.options(env.ci)
}
\arguments{
\item{env.ci}{Proportion of environmental variance used to define mature survival rate per generation
with a beta distribution. This number must be in between 0 and 1, but cannot be equal to 0 or 1.}
}
\value{
All values allowed and not allowed.
}
\description{
Given an env.ci value, we identified allowed means between 0 and 1.
}
