% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muvar.calc.R
\name{muvar.calc}
\alias{muvar.calc}
\title{Mature survival helper calculation.}
\usage{
muvar.calc(alpha, beta)
}
\arguments{
\item{alpha}{See alphabeta.calc.}

\item{beta}{See alphabeta.calc.}
}
\value{
List with the parameters mu and var.
}
\description{
This function can be used to convert alpha and beta back to mu and var.
}
