% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one.iter.f.choosy.R
\name{one.iter.f.choosy}
\alias{one.iter.f.choosy}
\title{Simulate a single generation}
\usage{
one.iter.f.choosy(
  ct.vec,
  aii.vec,
  as.matur,
  as.msurv.e.set,
  d,
  gnum.base,
  s,
  b,
  cc,
  mc,
  density.type = "all",
  mate.lazy = FALSE
)
}
\arguments{
\item{ct.vec}{Vector of current populations composition.}

\item{aii.vec}{The probability of an immature individual surviving (cytotype-specific).}

\item{as.matur}{Probability of maturation from immature to mature stage (list).}

\item{as.msurv.e.set}{Probability of an mature individual surviving (list).}

\item{d}{Strength of density dependency on gamete production.}

\item{gnum.base}{List of three numeric values, represent the base level mean
number of gametes per individual per cytotype. We defined gametes per
generation by a poisson distribution where lambda is equal to the this value
times the exponential of the product of strength of density dependency (ie. d)
and the sum of all individuals at time t or all individuals of a certain cytotype
at time t. Sum type is set by density type.}

\item{s}{Selfing rate.}

\item{b}{Proportion of unreduced gamete formed by each diploid and tetraploid individual.}

\item{cc}{Proportion of 3n gamete formation from each triploid individual.}

\item{mc}{Strength of mating choice.}

\item{density.type}{"all" sets gamete per generation sum as the sum of all
individuals at time t. "like-cytotype" sets gamete per generation sum as
the sum of both immature and mature individuals of the cytotype at time t.
Defaults to "all".}

\item{mate.lazy}{Set to be FALSE, unless set to TRUE. FALSE prevents selfing from occurring during outcrossing. However, this increases the computational time by 31x!}
}
\value{
List of 9, with 1:6 representing the number of individuals of each cytotype
and at both stages at time t + 1. Items 7:9 are the number of gametes sampled for
2x, 3x, and 4x individuals at time t.
}
\description{
Defined in Gaynor et al. Preprint, here we define a mixed-cytotypes population composition at time t + 1
based on the composition at time t and parameters related to reproduction, survival, and maturation.
}
