% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{stability}
\alias{stability}
\title{Stability metrics calculator following Ives et al. 2003}
\usage{
stability(B, sigma)
}
\arguments{
\item{B}{Matrix of p x p (p = number of species), where \eqn{b_{ij}} gives the effect of the abundance of species j on per capita growth rate of species.
This can be calculated with the \code{mars.cls()} function.}

\item{sigma}{Matrix of p x p, environmental noise variance-covariance matrix. This can be calculated with the \code{mars.cls()} function.}
}
\value{
Resulting list which includes:
\itemize{
\item \strong{var.prop}, at stationarity the var.prop is the variance proportion attributable to environmental noise. The smaller the values, the more stable the dynamics.
\item \strong{mean.return.time} & \strong{var.return.time}, rate at which the transition distribution converges back to the stationary distribution. The less time it takes to return to the stationary distribution, the more stable the population.
\item \strong{reactivity}, measures reaction to perturbations or the distance away from stationary a system moves in response to a disturbance. Again, smaller is better in terms of stability.
\item \strong{sp.contribs}, squared eigenvalue representing the characteristic return rate of the variance of the transition distribution of the estimated MARS(1) Markov Process.
}
}
\description{
Based on Ives et al. 2003, we calculate four stability  metrics based on the estimates of the matrix of ecological interactions
B and the estimated variance-covariance matrix of environmental noise 'Sigma'.  These two quantities can be read from the output of the function \code{mars.cls()}.
}
\references{
Ives, A. R., Dennis, B., Cottingham, K. L., Carpenter, S. R. (2003). Estimating community stability and ecological interactions frm time-series data. \emph{Ecological Monographs}, 72(2): 301 - 330.
}
