% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.shape.R
\name{surv.shape}
\alias{surv.shape}
\title{Survival shape -  Calculates the alpha and beta value for survival beta distributions}
\usage{
surv.shape(env.ci, raw.means)
}
\arguments{
\item{env.ci}{Proportion of environmental variance used to define mature survival rate per generation.
Must be a single integer greater than or equal to 0 and less than 1.}

\item{raw.means}{Mean probability of mature survival or immature survival. Must be a list of three integer between 0 and 1.}
}
\value{
Matrix of alpha and beta parameters where matrix\link{1,i} is the alpha value for the ith mean, and matrix\link{2,i}
is the beta value for the ith mean.
}
\description{
Calculates the shape parameters needed to sample a beta distribution
for both immature survival and mature survival
}
