% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.option.R
\name{var.option}
\alias{var.option}
\title{Mature survival variance calculation.}
\usage{
var.option(env.ci, mu)
}
\arguments{
\item{env.ci}{Proportion of environmental variance used to define mature survival rate per generation
with a beta distribution. This number must be in between 0 and 1, but cannot be equal to 0 or 1.}

\item{mu}{Mean probability of mature survival (list).}
}
\value{
Calculated variance.
}
\description{
Calculates variance based on env.ci, the proportion of env. variance, and mu, the mean probability of mature survival.
}
