% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.one.rep.R
\name{write.one.rep}
\alias{write.one.rep}
\title{Processing function to subset a single replicate from a model set.}
\usage{
write.one.rep(set, setname)
}
\arguments{
\item{set}{List containing paths to simulations wth replicates. This can be made with the function \code{list.files(path = "", full.names = TRUE)}.}

\item{setname}{Name for output folder.}
}
\value{
\code{write.one.rep()} creates a folder "output/one_rep/{setname}" where a csv file for each replicate is made.
}
\description{
Subset a single replicate from every model set.
}
