% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_one_rep.R
\name{write_one_rep}
\alias{write_one_rep}
\title{Processing function}
\usage{
write_one_rep(set, setname)
}
\arguments{
\item{set}{List of models}

\item{setname}{}
}
\value{
\code{write_one_rep()} creates a folder "output/one_rep/{setname}" where a csv file for each replicate is made.
}
\description{
Subset a single replicate from every model set.
}
