% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabeta.calc.R
\name{alphabeta.calc}
\alias{alphabeta.calc}
\title{Mature survival - Calculates alpha and beta from mean and variance.}
\usage{
alphabeta.calc(mu, var)
}
\arguments{
\item{mu}{Sample mean. Must be a single integer between 0 and 1.}

\item{var}{Derived variance based on var.option. Must be a single integer between 0 and 1.}
}
\value{
List with the parameters alpha and beta.
}
\description{
Calculates the alpha and beta parameters based on mu, here the mean probability of mature survival,
and s calculated variance.
}
