% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytotype_repro_mate.R
\name{cytotype_repro_mate}
\alias{cytotype_repro_mate}
\title{Reproduction function}
\usage{
cytotype_repro_mate(cgen, b, cc, gnum.vec, s, mc)
}
\arguments{
\item{cgen}{Number of mature individuals of each cytotype in the current generation.}

\item{b}{Proportion of unreduced gamete formed by each diploid and tetraploid individual.}

\item{cc}{Proportion of 3n gamete formation from each triploid individual.}

\item{gnum.vec}{The mean number of gametes produced by each individual (cytotype-specific).}

\item{s}{Selfing rate.}

\item{mc}{Strength of mating choice.}
}
\value{
Output is a generation log including number of 2x offspring, 3x offspring, 4x offspring, gametes sampled per 2x individual, gametes sampled per 3x individual, and gametes sampled per 4x individual.
}
\description{
This is a function to generate the number of offspring produced per generation.
Assumes a mixed mating system where both selfing and outcrossing occur.
Also allows group-based assortative mating (see Otto et al. 2008).
}
\examples{
\donttest{
extra.choosy <- cytotype_repro_mate(cgen = c(100,100,100),
                                    b = 0.02,
                                    cc = 0.052,
                                    gnum.vec = c(10, 10, 10),
                                    s = 0.02,
                                    mc = 1)
random.mating <- cytotype_repro_mate(cgen = c(100,100,100),
                                     b = 0.02,
                                     c = 0.052,
                                     gnum.vec = c(10, 10, 10),
                                     s = 0.02,
                                     mc = 0)
}

}
\references{
Otto, S. P., Servedio, M. R., Nuismer, S. L. (2008). Frequency-dependent selection and the evolution of assortative mating. \emph{Genetics}, 179(4) 2091 - 2112. doi: 10.1534/genetics.107.084418
}
